+++ Package.swift
@@ -1,0 +1,22 @@
+// swift-tools-version:5.1
+import PackageDescription
+
+let package = Package(
+    name: "XCDYouTubeKit",
+    platforms: [
+        .iOS(.v8),
+        .tvOS(.v9),
+        .macOS(.v10_10)
+    ],
+    products: [
+        .library(name: "XCDYouTubeKit" , targets: ["XCDYouTubeKit"])
+    ],
+    targets: [
+        .target(
+            name: "XCDYouTubeKit",
+            path: ".",
+            sources: ["XCDYouTubeKit"],
+            publicHeadersPath: "XCDYouTubeKit"
+        )
+    ]
+)