+++ XCDYouTubeKit/XCDYouTubeVideoOperation.m
@@ -25,6 +25,8 @@ typedef NS_ENUM(NSUInteger, XCDYouTubeRequestType) {
 @interface XCDYouTubeVideoOperation ()
 @property (atomic, copy, readonly) NSString *videoIdentifier;
 @property (atomic, copy, readonly) NSString *languageIdentifier;
+@property (atomic, strong, readonly) NSArray <NSHTTPCookie *> *cookies;
+@property (atomic, strong, readonly) NSArray <NSString *> *customPatterns;
 
 @property (atomic, assign) NSInteger requestCount;
 @property (atomic, assign) XCDYouTubeRequestType requestType;
@@ -52,7 +54,7 @@ typedef NS_ENUM(NSUInteger, XCDYouTubeRequestType) {
 
 static NSError *YouTubeError(NSError *error, NSSet *regionsAllowed, NSString *languageIdentifier)
 {
-	if (error.code == XCDYouTubeErrorRestrictedPlayback && regionsAllowed.count > 0)
+	if (error.code == XCDYouTubeErrorNoStreamAvailable && regionsAllowed.count > 0)
 	{
 		NSLocale *locale = [NSLocale localeWithLocaleIdentifier:languageIdentifier];
 		NSMutableSet *allowedCountries = [NSMutableSet new];
@@ -71,12 +73,15 @@ static NSError *YouTubeError(NSError *error, NSSet *regionsAllowed, NSString *la
 	}
 }
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wobjc-designated-initializers"
 - (instancetype) init
 {
-	@throw [NSException exceptionWithName:NSGenericException reason:@"Use the `initWithVideoIdentifier:languageIdentifier:` method instead." userInfo:nil];
+	@throw [NSException exceptionWithName:NSGenericException reason:@"Use the `initWithVideoIdentifier:cookies:languageIdentifier:` method instead." userInfo:nil];
 } // LCOV_EXCL_LINE
+#pragma clang diagnostic pop
 
-- (instancetype) initWithVideoIdentifier:(NSString *)videoIdentifier languageIdentifier:(NSString *)languageIdentifier
+- (instancetype) initWithVideoIdentifier:(NSString *)videoIdentifier languageIdentifier:(NSString *)languageIdentifier cookies:(NSArray<NSHTTPCookie *> *)cookies customPatterns:(NSArray<NSString *> *)customPatterns
 {
 	if (!(self = [super init]))
 		return nil; // LCOV_EXCL_LINE
@@ -85,10 +90,25 @@ static NSError *YouTubeError(NSError *error, NSSet *regionsAllowed, NSString *la
 	_languageIdentifier = languageIdentifier ?: @"en";
 	
 	_session = [NSURLSession sessionWithConfiguration:[NSURLSessionConfiguration ephemeralSessionConfiguration]];
+	_cookies = [cookies copy];
+	_customPatterns = [customPatterns copy];
 	
+	for (NSHTTPCookie *cookie in _cookies) {
+		[_session.configuration.HTTPCookieStorage setCookie:cookie];
+	}
 	_operationStartSemaphore = dispatch_semaphore_create(0);
 	
 	return self;
+}
+
+- (instancetype) initWithVideoIdentifier:(NSString *)videoIdentifier languageIdentifier:(NSString *)languageIdentifier
+{
+	return [self initWithVideoIdentifier:videoIdentifier languageIdentifier:languageIdentifier cookies:nil customPatterns:nil];
+}
+
+- (instancetype) initWithVideoIdentifier:(NSString *)videoIdentifier languageIdentifier:(NSString *)languageIdentifier cookies:(NSArray<NSHTTPCookie *> *)cookies
+{
+	return [self initWithVideoIdentifier:videoIdentifier languageIdentifier:languageIdentifier cookies:cookies customPatterns:nil];
 }
 
 #pragma mark - Requests
@@ -139,6 +159,7 @@ static NSError *YouTubeError(NSError *error, NSSet *regionsAllowed, NSString *la
 	
 	NSMutableURLRequest *request = [[NSMutableURLRequest alloc] initWithURL:url cachePolicy:NSURLRequestUseProtocolCachePolicy timeoutInterval:10];
 	[request setValue:self.languageIdentifier forHTTPHeaderField:@"Accept-Language"];
+	[request setValue:[NSString stringWithFormat:@"https://youtube.com/watch?v=%@", self.videoIdentifier] forHTTPHeaderField:@"Referer"];
 	
 	self.dataTask = [self.session dataTaskWithRequest:request completionHandler:^(NSData *data, NSURLResponse *response, NSError *error)
 	{
@@ -195,8 +216,8 @@ static NSError *YouTubeError(NSError *error, NSSet *regionsAllowed, NSString *la
 		return;
 	}
 	
-	NSDictionary *userInfo = @{ NSLocalizedDescriptionKey: connectionError.localizedDescription,
-	                            NSUnderlyingErrorKey: connectionError };
+	NSDictionary *userInfo = @{	NSLocalizedDescriptionKey: connectionError.localizedDescription,
+								NSUnderlyingErrorKey: connectionError };
 	self.lastError = [NSError errorWithDomain:XCDYouTubeVideoErrorDomain code:XCDYouTubeErrorNetwork userInfo:userInfo];
 	
 	[self startNextRequest];
@@ -234,7 +255,7 @@ static NSError *YouTubeError(NSError *error, NSSet *regionsAllowed, NSString *la
 		else
 		{
 			self.lastError = error;
-			if (error.code > 0)
+			if (error.userInfo[NSLocalizedDescriptionKey])
 				self.youTubeError = error;
 			
 			[self startNextRequest];
@@ -286,12 +307,12 @@ static NSError *YouTubeError(NSError *error, NSSet *regionsAllowed, NSString *la
 {
 	XCDYouTubeLogDebug(@"Handling JavaScript player response");
 	
-	self.playerScript = [[XCDYouTubePlayerScript alloc] initWithString:script];
+	self.playerScript = [[XCDYouTubePlayerScript alloc] initWithString:script customPatterns:self.customPatterns];
 	
-	if (self.webpage.isAgeRestricted)
+	if (self.webpage.isAgeRestricted && self.cookies.count == 0)
 	{
 		NSString *eurl = [@"https://youtube.googleapis.com/v/" stringByAppendingString:self.videoIdentifier];
-		NSString *sts = [(NSObject *)self.embedWebpage.playerConfiguration[@"sts"] description] ?: [(NSObject *)self.webpage.playerConfiguration[@"sts"] description] ?: @"";
+		NSString *sts = self.embedWebpage.sts ?: self.webpage.sts ?: @"";
 		NSDictionary *query = @{ @"video_id": self.videoIdentifier, @"hl": self.languageIdentifier, @"eurl": eurl, @"sts": sts};
 		NSString *queryString = XCDQueryStringWithDictionary(query);
 		NSURL *videoInfoURL = [NSURL URLWithString:[@"https://www.youtube.com/get_video_info?" stringByAppendingString:queryString]];